+++ src/main/java/com/anthonyhilyard/iceberg/Loader.java
@@ -41,6 +41,5 @@ public class Loader
 	@SubscribeEvent
 	public void onCommonSetup(FMLCommonSetupEvent event)
 	{
-
 	}
 }