+++ src/main/java/com/anthonyhilyard/iceberg/util/Selectors.java
@@ -127,6 +127,12 @@ public class Selectors
 			return validateSelector(value.substring(1));
 		}
 
+		// If this is a wildcard selector, it is valid.
+		if (value.contentEquals("*"))
+		{
+			return true;
+		}
+
 		// This is a tag, which should be a resource location.
 		if (value.startsWith("$"))
 		{